\name{energystuff-package}
\alias{energystuff-package}
\alias{energystuff}
\docType{package}
\title{
\packageTitle{energystuff}
}
\description{
\packageDescription{energystuff}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{energystuff}
\packageIndices{energystuff}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{energystuff}

Maintainer: \packageMaintainer{energystuff}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
