% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energystuff.R
\docType{package}
\name{energystuff}
\alias{energystuff}
\title{energystuff: a package for energy modeling and monitoring}
\description{
The energystuff package serves as a repository for various R functions that perform regression fitting and hypothesis testing.
}
\section{Hybrid energy monitoring}{

\code{hybridmonitor.single} function performs hypothesis testing for a single utility account.
\code{hybridmonitor.multiple} function equivalently performs hypothesis testing for multiple utility accounts by stochastically constructing a posterior predictive distribution of a test statistic, against which the test data are compared.
}

