% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridmonitor_multiple.R
\name{hybridmonitor.multiple}
\alias{hybridmonitor.multiple}
\title{Run hybrid monitoring procedure for multiple accounts}
\usage{
hybridmonitor.multiple(
  y,
  y_Kp1,
  m_idx = NULL,
  method,
  control = list(),
  ncores = NULL,
  verbose = FALSE
)
}
\arguments{
\item{y}{an array (12 x K x T) of measured energy usage to be used as the response variable.}

\item{y_Kp1}{a matrix of the last cycle of energy usage to be monitored.}

\item{m_idx}{(Optional) a vector of integers (1-12) that indicate the months to be tested. Note that 1 does not represent January. This completely depends on the user's definition of a \emph{cycle}. If not provided, it will by default run for all 12 months (i.e., full cycle).}

\item{method}{(Optional) a vector of characters ("bonferroni", "holm", "hochberg", "hommel", "BH", "BY") indicating the selection of multiplicity corrections. If not provided, the multiplicity correction will default to Bonferroni's method.}

\item{control}{(Optional) a list of parameters for the \href{https://en.wikipedia.org/wiki/Monte_Carlo_method}{Monte Carlo sampling} of the posterior predictive distribution of \emph{p}-values, as well as the \emph{p}-value of the test data.}

\item{ncores}{(Optional)}
}
\value{
a dataframe with the result of the hypothesis testing, empirical distribution of the \emph{p}-values, observed \emph{p}-value
}
\description{
This is a function that run hybrid monitoring algorithm for multiple accounts
}
\examples{
\dontrun{
fit <- hybridmonitor.multiple(y, y_Kp1)
}
}
\seealso{
\code{\link{hybridmonitor.single}} for single-account monitoring
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
