% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridmonitor_single.R
\name{hybridmonitor.single}
\alias{hybridmonitor.single}
\title{Run hybrid monitoring procedure for a single account}
\usage{
hybridmonitor.single(
  y,
  y_Kp1,
  control = list(),
  ncores = NULL,
  verbose = FALSE
)
}
\arguments{
\item{y}{a vector of measured energy usage to be used as the response variable}
}
\value{
a dataframe with the result of the hypothesis testing, empirical distribution of the p-values, observed p-value
}
\description{
This is a function that run hybrid monitoring algorithm for a single account
}
\examples{
\dontrun{
fit <- hybridmonitor.single(y, control=list(C=rep(5,12)))
}
}
\author{
Daeyoung Lim, \email{daeyoung.lim@uconn.edu}
}
